% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_conf_lvl}
\alias{check_conf_lvl}
\title{INTERNAL: Check if a confidence level is entered in decimal form}
\usage{
check_conf_lvl(level)
}
\arguments{
\item{level}{The confidence level in question}
}
\value{
Nothing, or an error message.
}
\description{
Used internally to detect if a user inputted a confidence level correctly (i.e., not in quotes
and as a decimal).
}
\examples{
check_conf_lvl(.95)
try(check_conf_lvl("95\%"))
try(check_conf_lvl(95))
}
\keyword{internal}
