% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_test}
\alias{check_test}
\title{INTERNAL: Check if an error will occur later}
\usage{
check_test(code)
}
\arguments{
\item{code}{The code to try to run}
}
\value{
Nothing, but may print an error message.
}
\description{
Used internally to check if an error might occur later. We catch the error here to provide
a more helpful message to users based on the most common causes of errors.
}
\examples{
library(stats)

check_test(head(mtcars))
try(check_test(head(Mtcars)))

check_test(lm(wt~drat, data = mtcars))
try(check_test(lm(Wt~drat, data = mtcars)))
}
\keyword{internal}
