% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finalize_plot}
\alias{finalize_plot}
\title{Add default theme options to plots}
\usage{
finalize_plot(plot)
}
\arguments{
\item{plot}{A plot (to be piped into \code{ggformula::gf_theme})}
}
\value{
A \code{ggplot} object..
}
\description{
\code{finalize_plot()} is used internally to complete the plot creation process. This function uses \code{ggformula}
to create a \code{ggplot} object and adds theme choices and size specifications.
}
