% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gvsu215-package.R
\docType{package}
\name{gvsu215-package}
\alias{gvsu215}
\alias{gvsu215-package}
\title{gvsu215: Pretty Output For Select 'mosaic' and 'ggformula' Functions}
\description{
A light wrapper to 'mosaic' and 'ggformula' enabling nice graphics in one line of code. Designed for the Grand Valley State University's Introductory Statistics course (STA 215).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ian-curtis/gvsu215}
  \item \url{https://ian-curtis.github.io/gvsu215/}
  \item Report bugs at \url{https://github.com/ian-curtis/gvsu215/issues}
}

}
\author{
\strong{Maintainer}: Ian Curtis \email{curtisi@mail.gvsu.edu} [copyright holder]

}
\keyword{internal}
