% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_mean.R
\name{infer_1mean_int}
\alias{infer_1mean_int}
\title{Create a summary table for a one-sample mean interval}
\usage{
infer_1mean_int(
  data,
  formula,
  digits = 3,
  null = 0,
  conf_lvl = 0.95,
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variable to run the test on, in formula syntax, \code{~var}.}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{null}{The value of the null hypothesis. Defaults to 0.}

\item{conf_lvl}{The confidence level of the interval, entered as a value between 0 and 1.
Defaults to 0.95.}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable. In an interactive environment, results are viewable immediately.
}
\description{
Create a summary table for a one-sample mean interval
}
\examples{
infer_1mean_int(mtcars, ~wt)
infer_1mean_int(mtcars, ~wt, conf_lvl = .9)
}
