% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_prop_mean.R
\name{infer_1prop}
\alias{infer_1prop}
\title{Create a summary table for a one-sample proportion interval}
\usage{
infer_1prop(
  data,
  formula,
  success = NULL,
  digits = 3,
  conf_lvl = 0.95,
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variable to run the test on, in formula syntax, \code{~var}.}

\item{success}{The data value that constitutes a "success".}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{conf_lvl}{The confidence level of the interval, entered as a value between 0 and 1.
Defaults to 0.95.}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable. In an interactive environment, results are viewable immediately.
}
\description{
Create a summary table for a one-sample proportion interval
}
\examples{
infer_1prop(mtcars, ~vs, success = 1)
infer_1prop(mtcars, ~vs, success = 1, conf_lvl = .9)
}
