% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_alt.R
\name{infer_2mean}
\alias{infer_2mean}
\title{All-in-one alternate two-sample means test}
\usage{
infer_2mean(
  data,
  formula,
  digits = 3,
  null = 0,
  conf_lvl = 0.95,
  conf_int = c("hide", "show"),
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variables to run the test on, in formula syntax, \code{var1 ~ var2}.}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{null}{The value of the null hypothesis. Defaults to 0.}

\item{conf_lvl}{The confidence level of the interval, entered as a value between 0 and 1.
Defaults to 0.95.}

\item{conf_int}{Should a confidence interval be provided in addition to the hypothesis test output?
Defaults to "hide" with the other option being "show".}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable. In interactive sessions, output is viewable immediately.
}
\description{
\code{infer_2mean()} is an alternative to \code{\link[=infer_2mean_test]{infer_2mean_test()}} and \code{\link[=infer_2mean_int]{infer_2mean_int()}}. Rather than have
hypothesis test and confidence interval output split into two separate functions, you can now do it
in one. For just a hypothesis test, do nothing different from \code{\link[=infer_2mean_test]{infer_2mean_test()}} (except change
the function name). For a confidence interval provided with that, use \code{conf_int = TRUE}.
}
\examples{
infer_2mean(mtcars, wt~vs)
infer_2mean(mtcars, wt~vs, conf_lvl = .9)
infer_2mean(mtcars, wt~vs, conf_lvl = .9, conf_int = "hide")
}
