% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_alt.R
\name{infer_2prop}
\alias{infer_2prop}
\title{All-in-one alternate two-sample proportion test}
\usage{
infer_2prop(
  data,
  formula,
  success,
  digits = 3,
  conf_lvl = 0.95,
  alternative = c("notequal", "greater", "less"),
  conf_int = c("hide", "show"),
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variables to run the test on, in formula syntax, \code{var1 ~ var2}.}

\item{success}{The data value that constitutes a "success".}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{conf_lvl}{The confidence level of the interval, entered as a value between 0 and 1.
Defaults to 0.95.}

\item{alternative}{The alternative hypothesis. Defaults to "notequal" (two sided p-value).
Other options include "greater" or "less". Use depends on your test.}

\item{conf_int}{Should a confidence interval be provided in addition to the hypothesis test output?
Defaults to "hide" with the other option being "show".}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable. In interactive sessions, output is viewable immediately.
}
\description{
\code{infer_2prop()} is an alternative to \code{\link[=infer_2prop_test]{infer_2prop_test()}} and \code{\link[=infer_2prop_int]{infer_2prop_int()}}. Rather than have
hypothesis test and confidence interval output split into two separate functions, you can now do it
in one. For just a hypothesis test, do nothing different from \code{\link[=infer_2prop_test]{infer_2prop_test()}} (except change
the function name). For a confidence interval provided with that, use \code{conf_int = "show"}.
}
\examples{
infer_2prop(mtcars, vs~am, success = 1)
infer_2prop(mtcars, vs~am, success = 1, conf_lvl = .9, digits = 4)
infer_2prop(mtcars, vs~am, success = 1, conf_lvl = .9, digits = 5, conf_int = "show")
}
