% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_other.R
\name{infer_anova}
\alias{infer_anova}
\title{Create a summary table}
\usage{
infer_anova(data, formula, digits = 3, caption = NULL)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variables to run the test on, in formula syntax, \code{var1 ~ var2}.}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable. If in an interactive session, results are viewable immediately.
}
\description{
Create a summary table
}
\examples{
infer_anova(mtcars, drat~gear)
infer_anova(mtcars, drat~gear, digits = 4)
}
