% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_other.R
\name{infer_chisq}
\alias{infer_chisq}
\title{Create a summary table for a chi-squared test}
\usage{
infer_chisq(
  data,
  formula,
  type = c("test", "expected", "observed"),
  digits = 3,
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variable to run the test on, in formula syntax, \code{~var}.}

\item{type}{The kind of output to receive. Valid options are either "test" (the default), "expected",
or "observed".}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable.
}
\description{
\code{infer_chisq()} creates a tidy summary table for results of a chi-squared test. Alternatively, you can
select to see expected counts (\code{type = "expected"}) or observed counts (\code{type = "observed"}).
}
\examples{
suppressWarnings(infer_chisq(mtcars, cyl~gear))
suppressWarnings(infer_chisq(mtcars, cyl~gear, type = "expected"))
suppressWarnings(infer_chisq(mtcars, cyl~gear, type = "observed"))
}
