% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_mean.R
\name{infer_paired}
\alias{infer_paired}
\title{Create a summary table for a paired means test}
\usage{
infer_paired(
  data,
  var1,
  var2,
  digits = 3,
  null = 0,
  conf_lvl = 0.95,
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{var1}{The first variable of the pair, entered in formula syntax \code{var1}.}

\item{var2}{The second variable of the pair, entered in formula syntax \code{var2}.}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{null}{The value of the null hypothesis. Defaults to 0.}

\item{conf_lvl}{The confidence level of the interval, entered as a value between 0 and 1.
Defaults to 0.95.}

\item{caption}{An override to the table caption. A sensible default is provided.}
}
\value{
An object of class flextable. In an interactive environment, results are viewable immediately.
}
\description{
Create a summary table for a paired means test
}
\examples{
infer_paired(mtcars, var1 = ~drat, var2 = ~wt)
infer_paired(mtcars, var1 = ~drat, var2 = ~wt, conf_lvl = 0.9)
}
