% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_other.R
\name{infer_reg}
\alias{infer_reg}
\title{Create a summary table for a linear regression test}
\usage{
infer_reg(data, formula, digits = 3, caption = NULL, reduced = TRUE)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{The variables to run the test on, in formula syntax. Passed on to \code{\link[stats:lm]{stats::lm()}}.}

\item{digits}{The number of digits to round table values to. Defaults to 3.}

\item{caption}{An override to the table caption. A sensible default is provided.}

\item{reduced}{Should a simple table be created (i.e., removal of the \code{t} and \code{p-value} columns)?
Defaults to TRUE.}
}
\value{
An object of class flextable. In an interactive session, results are viewable immediately.
}
\description{
\code{infer_reg()} creates a small table summarizing a linear regression test. The traditional columns
of \code{t} and \code{p-value} are removed by default but a full table can be accessed with simple = FALSE.
}
\examples{
infer_reg(mtcars, drat~wt)
infer_reg(mtcars, drat~wt, reduced = FALSE)
infer_reg(mtcars, drat~wt, digits = 4)
infer_reg(mtcars, drat~wt + qsec)
}
