% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{one_way}
\alias{one_way}
\title{Create a one way data table}
\usage{
one_way(data, formula, digits = 3, caption = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variable to tabulate. Should be given in formula notation, \code{~var}.}

\item{digits}{The number of digits to round to. Defaults to 3.}

\item{caption}{An optional caption for the table. Defaults to \code{NULL}, but a sensible default is
provided within the function.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
\code{one_way()} creates a table giving counts and percentages on one variable.
}
\examples{
one_way(mtcars, ~gear)
one_way(mtcars, ~gear, digits = 2)

# Will give an error
try(one_way(mtcars, ~Gear))
}
