% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{pctile}
\alias{pctile}
\title{Create percentile summaries}
\usage{
pctile(
  data,
  formula,
  digits = 3,
  probs = c(0, 0.25, 0.5, 0.75, 1),
  caption = NULL,
  na_rm = FALSE
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{Variables to summarize given in formula notation: \code{~var1} or \code{var1~var2}.
\code{var1} should be numeric and \code{var2}, when supplied, should be a grouping variable.}

\item{digits}{The number if digits to round to. Defaults to 3.}

\item{probs}{A vector of percentiles to compute. Each value must be between 0 and 1, inclusive.
Defaults to 0, 25\%, 50\%, 75\%, and 100\%.}

\item{caption}{An override for the table caption. A sensible default is given.}

\item{na_rm}{Should missing values be removed? Defaults to FALSE.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
\code{pctile()} is a wrapper around \code{mosaic::quantile()} and creates a tidy table of data values
at the given percentiles.
}
\examples{
pctile(mtcars, ~wt)
pctile(mtcars, ~wt, probs = c(.17, .3, .5, .7, .9, 1), na_rm = TRUE)
pctile(mtcars, wt~cyl, na_rm = TRUE)

try(pctile(mtcars, ~wt, probs = c(25, 50, 75, 100)))
}
