% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bar}
\alias{plot_bar}
\title{Create a simple bar plot}
\usage{
plot_bar(
  data,
  formula,
  type = c("percent", "count"),
  fill = "#0032A0",
  title = NULL,
  na_rm = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variable to tabulate. Should be given in formula notation, \code{~var}.}

\item{type}{The type of plot to create. Valid options are "percent" (the default) or "count".}

\item{fill}{The fill of the plot. Valid options are a character color (for one variable plots) or
a variable given in formula notation (\code{~var}), used to create a grouped bar plot.}

\item{title}{An override for the title of the plot. A sensible default is provided.}

\item{na_rm}{Should missing values be removed? Defaults to FALSE.}

\item{...}{Extra title arguments passed on to \code{\link[mosaic:gf_labs]{mosaic::gf_labs()}} (which feeds to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}).}
}
\value{
A ggplot object. In an interactive session, results are viewable immediately.
}
\description{
Create a simple bar plot
}
\examples{
plot_bar(mtcars, ~cyl)
plot_bar(mtcars, ~cyl, type = "count")
plot_bar(mtcars, ~cyl, type = "percent", fill = "yellowgreen")

plot_bar(mtcars, ~cyl, fill = ~gear)
plot_bar(mtcars, ~cyl, type = "count", fill = ~gear)
}
