% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hist}
\alias{plot_hist}
\title{Create a simple histogram}
\usage{
plot_hist(
  data,
  formula,
  fill = "#0032A0",
  binwidth = NULL,
  group = NULL,
  group_cols = 1,
  title = NULL,
  na_rm = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variable to tabulate. Should be given in formula notation, \code{~var}.}

\item{fill}{The fill color of the bins, entered as a character.}

\item{binwidth}{The width of the bins, entered as a number.}

\item{group}{A grouping (faceting) variable entered in formula syntax, \code{~group_var}.}

\item{group_cols}{The number of columns to make in a grouped (faceted) plot. Defaults to
1 (stacked vertical plots).}

\item{title}{An override for the title of the plot. A sensible default is provided.}

\item{na_rm}{Should missing values be removed? Defaults to FALSE.}

\item{...}{Extra title arguments passed on to \code{\link[ggformula:gf_aux]{ggformula::gf_labs()}} (which feeds to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}).}
}
\value{
A ggplot object. In an interactive environment, results are viewable immediately.
}
\description{
\code{plot_hist()} builds a histogram on one variable alone or grouped by a single, categorical grouping
variable. The number of grouped plot columns can be adjusted based on the levels of the
grouping variable.
}
\examples{
plot_hist(mtcars, ~drat)
plot_hist(mtcars, ~drat, binwidth = 0.05)
plot_hist(mtcars, ~drat, binwidth = 0.05, fill = "red")

plot_hist(mtcars, ~drat, binwidth = 0.05, group = ~cyl)
plot_hist(mtcars, ~drat, binwidth = 0.05, group = ~cyl, group_cols = 2)
}
