% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sig_round}
\alias{sig_round}
\title{INTERNAL: Round a number using standard rounding or sigfigs}
\usage{
sig_round(number, digits)
}
\arguments{
\item{number}{The number to be rounded.}

\item{digits}{The number of digits to round to. If the number is between -1 and 1, the value is given to
\code{signif()} and rounds to number of non-zero digits. Otherwise, value is given to \code{round()}. If the
rounded value is shorter in length than the number of digits provided, zeroes are padded to the end.}
}
\value{
A character of the rounded number.
}
\description{
INTERNAL: Round a number using standard rounding or sigfigs
}
\examples{
sig_round(0.00023556, 3)
sig_round(23.1233644, 3)
sig_round(0.331, 4)
sig_round(4, 5)
}
