% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_1var}
\alias{tbl_1var}
\title{Create a one way data table}
\usage{
tbl_1var(
  data,
  formula,
  digits = 3,
  caption = NULL,
  with_prop = c("no", "yes"),
  na_rm = FALSE
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variable to tabulate. Should be given in formula notation, \code{~var}.}

\item{digits}{The number of digits to round to. Defaults to 3.}

\item{caption}{An optional caption for the table. Defaults to \code{NULL}, but a sensible default is
provided within the function.}

\item{with_prop}{Should proportions be supplied with instead of percents? Defaults to "no",
could also be "yes".}

\item{na_rm}{Should missing values be removed? Defaults to FALSE.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
\code{tbl_1var()} creates a table giving counts and percentages on one variable.
}
\examples{
tbl_1var(mtcars, ~gear)
tbl_1var(mtcars, ~gear, digits = 2)

# Will give an error
try(tbl_1var(mtcars, ~Gear))
}
