% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tbl_2var}
\alias{tbl_2var}
\title{Create a two way data table}
\usage{
tbl_2var(
  data,
  formula,
  row_pct = c("hide", "show"),
  digits = 3,
  caption = NULL,
  na_rm = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variables to tabulate. Should be given in formula notation \code{var1~var2}. Changing
the order of the variables will swap the table axes.}

\item{row_pct}{Should row percents be included in each cell? Defaults to "hide" with the only
other possible value being "show".}

\item{digits}{The number of digits to round to. Defaults to 3.}

\item{caption}{An optional caption for the table. Defaults to \code{NULL}, but a sensible default is
provided within the function.}

\item{na_rm}{Should missing values be removed? Defaults to TRUE.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
Create a two way data table
}
\examples{
tbl_2var(mtcars, cyl~gear)
tbl_2var(mtcars, cyl~gear, row_pct = "show")
tbl_2var(mtcars, cyl~gear, caption = "This is the new caption")

# Will give an error
try(tbl_2var(mtcars, Cyl~Gear))
}
