% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{tbl_corr}
\alias{tbl_corr}
\title{Create a simple correlation table}
\usage{
tbl_corr(data, formula, digits = 3, caption = NULL, na_rm = FALSE)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{Two variables given in formula notation: \code{var1~var2}.}

\item{digits}{The number if digits to round to. Defaults to 3.}

\item{caption}{An override for the table caption. A sensible default is given.}

\item{na_rm}{Should missing values be removed? Defaults to FALSE.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
Create a simple correlation table
}
\examples{
# Not removing NAs is not recommended
tbl_corr(mtcars, wt~qsec)

tbl_corr(mtcars, wt~qsec, na_rm = TRUE)
}
