% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{tbl_pctile}
\alias{tbl_pctile}
\title{Create percentile summaries}
\usage{
tbl_pctile(
  data,
  formula,
  digits = 3,
  probs = c(0, 0.25, 0.5, 0.75, 1),
  caption = NULL
)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{Variables to summarize given in formula notation: \code{~var1} or \code{var1~var2}.
\code{var1} should be numeric and \code{var2}, when supplied, should be a grouping variable.}

\item{digits}{The number if digits to round to. Defaults to 3.}

\item{probs}{A vector of percentiles to compute. Each value must be between 0 and 1, inclusive.
Defaults to 0, 25\%, 50\%, 75\%, and 100\%.}

\item{caption}{An override for the table caption. A sensible default is given.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
\code{tbl_pctile()} is a wrapper around \code{mosaic::quantile()} and creates a tidy table of data values
at the given percentiles.
}
\examples{
tbl_pctile(mtcars, ~wt)
tbl_pctile(mtcars, ~wt, probs = c(.17, .3, .5, .7, .9, 1))
tbl_pctile(mtcars, wt~cyl, digits = 4)

try(tbl_pctile(mtcars, ~wt, probs = c(25, 50, 75, 100)))
}
