% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bar}
\alias{plot_bar}
\title{Create a simple bar plot}
\usage{
plot_bar(
  data,
  formula,
  type = c("percent", "count"),
  fill = "#0032A0",
  fill_position = c("sbs", "stack"),
  orient = c("vertical", "horizontal"),
  dodge = 1,
  title = NULL,
  subtitle = c("show", "hide"),
  na_rm = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variable to tabulate. Should be given in formula notation, \code{~var}.}

\item{type}{The type of plot to create. Valid options are "percent" (the default) or "count".}

\item{fill}{The fill of the plot. Valid options are a character color (for one variable plots) or
a variable given in formula notation (\code{~var}), used to create a grouped bar plot.}

\item{fill_position}{The bar type for grouped plots. Either \code{"sbs"} for side-by-side bars or \code{"stack"} for stacked bars. \code{fill} must be a variable for this to go into effect.}

\item{orient}{The orientation for the plot (either "vertical", the default, or "horizontal"). As a
shortcut, "v" and "h" may be used.}

\item{dodge}{The number of rows to dodge the axis labels to should they be overlapping.}

\item{title}{An override for the title of the plot. A sensible default is provided.}

\item{subtitle}{A switch for hiding the default subtitle. One of "show" or "hide".}

\item{na_rm}{Should missing values be removed? Defaults to TRUE.}

\item{...}{Extra title arguments passed on to \code{\link[ggformula:gf_aux]{ggformula::gf_labs()}} (which feeds to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}).}
}
\value{
A ggplot object. In an interactive session, results are viewable immediately.
}
\description{
Create a simple bar plot
}
\examples{
plot_bar(mtcars, ~cyl)
plot_bar(mtcars, ~cyl, type = "count")
plot_bar(mtcars, ~cyl, type = "percent", fill = "yellowgreen")
plot_bar(mtcars, ~cyl, orient = "horizontal")
plot_bar(dplyr::starwars, ~hair_color, dodge = 2)
plot_bar(mtcars, ~cyl, subtitle = "hide")

plot_bar(mtcars, ~cyl, fill = ~gear)
plot_bar(mtcars, ~cyl, fill = ~gear, fill_position = "stack")
plot_bar(mtcars, ~cyl, type = "count", fill = ~gear)
}
