% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hist}
\alias{plot_hist}
\title{Create a simple histogram}
\usage{
plot_hist(
  data,
  formula,
  fill = "#0032A0",
  breaks = NULL,
  group = NULL,
  group_cols = 1,
  title = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{The variable to tabulate. Should be given in formula notation, \code{~var}.}

\item{fill}{The fill color of the bins, entered as a character.}

\item{breaks}{A vector of length 3 (start, stop, step) specifying how the x-scale should be broken up.
Make sure that (stop - start)/step is a whole number. See notes for more details.}

\item{group}{A grouping (faceting) variable entered in formula syntax, \code{~group_var}.}

\item{group_cols}{The number of columns to make in a grouped (faceted) plot. Defaults to
1 (stacked vertical plots).}

\item{title}{An override for the title of the plot. A sensible default is provided.}

\item{...}{Extra title arguments passed on to \code{\link[ggformula:gf_aux]{ggformula::gf_labs()}} (which feeds to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}).}
}
\value{
A ggplot object. In an interactive environment, results are viewable immediately.
}
\description{
\code{plot_hist()} builds a histogram on one variable alone or grouped by a single, categorical grouping
variable. The number of grouped plot columns can be adjusted based on the levels of the
grouping variable.

Default histograms will likely not show your data in the best way (the function will warn you of this
once per session). To remedy this, use the \code{breaks} argument. The values given in the vector should
all be numbers. The first number is the minimum value you want displayed, the second is the maximum
value to display, and the third is the "step" distance. (For those who are familiar with \code{ggplot2},
the first and second value are supplied to \code{limits} and a sequence of the three numbers are
supplied to \code{breaks} of \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}.) For best results, (stop - start) / step
should be a whole number.
}
\examples{
plot_hist(mtcars, ~drat)
plot_hist(mtcars, ~drat, breaks = c(1, 6, 0.5))
plot_hist(mtcars, ~drat, breaks = c(1, 6, 0.5), fill = "red")

plot_hist(mtcars, ~drat, group = ~cyl, breaks = c(2, 5, 0.25))
plot_hist(mtcars, ~drat, group = ~cyl, breaks = c(2, 5, 0.25), group_cols = 2)
}
