% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Create a simple scatterplot}
\usage{
plot_scatter(
  data,
  formula,
  fill = "#0032a0",
  title = NULL,
  legend_title = NULL,
  axis_lines = c("none", "both"),
  ls_line = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{Variables to build the plot on. Should be entered in formula notation, \code{var1~var2}.}

\item{fill}{The fill of the plot. Valid options are a character color (for standard scatterplots) or
a variable given in formula notation (\code{~var}), used to create a grouped scatterplot.}

\item{title}{An override for the title of the plot. A sensible default is provided.}

\item{legend_title}{The title of the lengend. Ignored in non-grouped plots. Default is the variable name.}

\item{axis_lines}{Should major axis lines appear on the plot? Valid options are "none" or "both.
Defaults to "none".}

\item{ls_line}{Should a least squares line (or lines) appear on the plot? Defaults to FALSE.}

\item{...}{Extra title arguments passed on to \code{\link[ggformula:gf_aux]{ggformula::gf_labs()}} (which feeds to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}).}
}
\value{
A ggplot object. In an interactive environment, results are viewable immediately.
}
\description{
\code{plot_scatter()} creates a pre-themed scatterplot on two variables, optionally grouped by a
third categorical variable.
}
\examples{
plot_scatter(mtcars, wt~drat)
plot_scatter(mtcars, wt~drat, fill = "red")
plot_scatter(mtcars, wt~drat, axis_lines = "both")
plot_scatter(mtcars, wt~drat, ls_line = TRUE)

plot_scatter(mtcars, wt~drat, fill = ~cyl)
plot_scatter(mtcars, wt~drat, fill = ~cyl, ls_line = TRUE)
plot_scatter(mtcars, wt~drat, fill = ~cyl, legend_title = "Cylinders")
}
