% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{tbl_num_sum}
\alias{tbl_num_sum}
\title{Create numerical summaries}
\usage{
tbl_num_sum(data, formula, digits = 3, na_rm = FALSE, caption = NULL)
}
\arguments{
\item{data}{A data frame (or tibble).}

\item{formula}{Variables to summarize given in formula notation: \code{~var1} or \code{var1~var2}.
\code{var1} should be numeric and \code{var2}, when supplied, should be a grouping variable.}

\item{digits}{The number if digits to round to. Defaults to 3.}

\item{na_rm}{Should missing values be removed? Defaults to FALSE.}

\item{caption}{An override for the table caption. A sensible default is given.}
}
\value{
An object of class flextable. If in an interactive session, the table will be
viewable immediately.
}
\description{
\code{tbl_num_sum()} is a wrapper around \code{mosaic::favstats()} and creates a tidy table of summary statistics
including the min, q1, median, q3, max, mean, standard deviation, n, and missing values.
}
\examples{
# not removing NAs is not recommended
tbl_num_sum(mtcars, ~wt)

tbl_num_sum(mtcars, ~wt, na_rm = TRUE)
tbl_num_sum(mtcars, ~wt, na_rm = TRUE, digits = 2, caption = "This is a table")
tbl_num_sum(mtcars, wt~cyl, na_rm = TRUE)
}
